+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # changelog
 
+## v0.3.1
+
 * optimize `AbsoluteInt` (with Two's complement implementation)
 * panic if call `AbsoluteInt` with the minimum value of integers (the absolute of the minimum value of integers is impossible to return)